src/requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astroid==3.3.9
asttokens==2.4.1
attrs==23.2.0
beir==2.1.0
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==7.1.2
cloudpathlib==0.18.1
coloredlogs==15.0.1
comet-ml==3.47.0
comm==0.2.2
confection==0.1.5
configobj==5.0.9
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.12
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
dulwich==0.22.1
elasticsearch==7.9.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entmax==1.3
evaluate==0.4.3
everett==3.1.0
executing==2.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numlet==2.0.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.20.1
openai==1.65.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyjnius==1.6.1
pylint==3.3.7
pypalettes==0.1.3
pyparsing==3.1.2
pyserini==0.44.0
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrec-eval-terrier==0.5.6
pytrec_eval==0.5
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rouge_score==0.1.2
rpds-py==0.20.0
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
sentry-sdk==2.10.0
setproctitle==1.3.3
setuptools==71.0.2
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.20.3
tomlkit==0.13.2
torch==2.6.0
torchtext==0.18.0
torchvision==0.21.0
torchviz==0.0.3
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.0
triton==3.2.0
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.4
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wget==3.2
wrapt==1.16.0
wurlitzer==3.1.1
xxhash==3.4.1
yarl==1.9.4
